% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{comingle_pprof}
\alias{comingle_pprof}
\title{Profile R and native code and record pprof samples.}
\usage{
comingle_pprof(expr, pprof = tempfile(), ...)
}
\arguments{
\item{expr}{An R expression to profile.}

\item{pprof}{Path to a file with pprof samples.
Also returned from the function.}

\item{...}{Passed on to \code{start_profiler()}.}
}
\value{
Path to a file with pprof samples.
}
\description{
Run R code and record pprof samples
from both R's and the native tracebacks.
This function uses \code{\link[=start_profiler]{start_profiler()}} and \code{\link[=stop_profiler]{stop_profiler()}}
to initiate and stop the data collection.
}
\examples{
# Returns a path to pprof samples.
comingle_pprof(replicate(1e2, sample.int(1e4)))
}
