% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{comingle_rprof}
\alias{comingle_rprof}
\title{Profile R code and record Rprof samples.}
\usage{
comingle_rprof(expr, rprof = tempfile(), ...)
}
\arguments{
\item{expr}{An R expression to profile.}

\item{rprof}{Path to a file with Rprof samples.
Also returned from the function.}

\item{...}{Passed on to \code{start_profiler()}.}
}
\value{
Path to a file with Rprof samples.
}
\description{
Run R code and record Rprof samples
from both R's and the native tracebacks.
This function uses \code{\link[=start_profiler]{start_profiler()}} and \code{\link[=stop_profiler]{stop_profiler()}}
to initiate and stop the data collection.
}
\examples{
# Returns a path to Rprof samples.
comingle_rprof(replicate(1e2, sample.int(1e4)))
}
