% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof.R
\name{find_pprof}
\alias{find_pprof}
\title{Find the pprof utility}
\usage{
find_pprof()
}
\description{
\code{pprof} is a tool for visualization and analysis of profiling data,
implemented in Go.
The \code{find_pprof()} function is a wrapper around \code{Sys.which()}.
It fails if no \code{pprof} executable can be found.
}
\section{Installation}{

Use the instructions on \code{pprof}'s
\href{https://github.com/google/pprof#building-pprof}{GitHub page}
to install it.  Make sure the \code{pprof} executable is on your system path.
}

