% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{get_profiler_traces}
\alias{get_profiler_traces}
\title{Parse profiler output}
\usage{
get_profiler_traces(path = "1.prof")
}
\arguments{
\item{path}{The path to the profiler output.}
}
\description{
Combines the profiler output obtained from \code{\link[=start_profiler]{start_profiler()}} into
a nested tibble.
}
