% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointprof.R
\name{joint_pprof}
\alias{joint_pprof}
\title{Profile an expression and show in pprof}
\usage{
joint_pprof(expr, ...)
}
\arguments{
\item{expr}{Expression to profile.}

\item{...}{Passed on to \code{\link[proffer:serve_pprof]{proffer::serve_pprof()}}.}
}
\description{
Profiles an expression with \code{\link[=start_profiler]{start_profiler()}} and \code{\link[=stop_profiler]{stop_profiler()}}
and visualizes results with \code{\link[proffer:serve_pprof]{proffer::serve_pprof()}}.
Drop-in replacement for \code{\link[proffer:pprof]{proffer::pprof()}} and \code{\link[profvis:profvis]{profvis::profvis()}}.
}
