% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler-package.R
\docType{package}
\name{jointprof-package}
\alias{jointprof}
\alias{jointprof-package}
\title{jointprof: Joint Profiling of Native and R Code}
\description{
Simultaneously collects profiling data at both R and native levels
(using 'Rprof' and 'gprofiler') and provides a unified view of the collected data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-prof/gprofiler}
  \item \url{http://r-prof.github.io/gprofiler}
  \item Report bugs at \url{https://github.com/r-prof/gprofiler/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item Google, Inc. (Bundled pprof sources, and code extracted from gperftools) [copyright holder]
}

}
