% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{start_profiler}
\alias{start_profiler}
\alias{stop_profiler}
\title{Starts and stops profiling}
\usage{
start_profiler(path = "1.prof")

stop_profiler()
}
\arguments{
\item{path}{Path to the output file. Other files based on this path
may be created.}
}
\description{
\code{start_profiler()} initiates profiling.

\code{stop_profiler()} terminates profiling. The results are available with
\code{\link[=get_profiler_traces]{get_profiler_traces()}}.
}
