% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_pprof}
\alias{vis_pprof}
\title{Visualize profile data with \code{pprof}.}
\usage{
vis_pprof(path, host = "localhost", port = NULL)
}
\arguments{
\item{path}{Character, path to a profile data file in
\href{https://github.com/google/pprof}{pprof} format.
Convert between formats using the
\href{https://github.com/r-prof/profile}{profile} package.}

\item{host}{Character, IP address or complete hostname.
Determines which computers can connect to the
local \href{https://github.com/google/pprof}{pprof} web server.
In most cases, the default \code{"localhost"} is fine.
To allow remote connections from other computers, use \code{"0.0.0.0"}.}

\item{port}{Character, port number to host the interactive displays.
If \code{NULL} (default) a random port between 49152 and 65355 is
selected uniformly at random.

A port number is a number between 0 and 65535 to identify a service
running on a computer. Ports 49152 to 65355 tend to be
more available than 1024 to 49151. Avoid ports 0 to 1023.}
}
\description{
\href{https://github.com/google/pprof}{pprof}
is a tool for visualization and analysis of profile data,
implemented in Go. See \url{https://r-prof.github.io/jointprof}
for installation instructions.

The \code{vis_pprof()} function launches a local web server
to display profile data interactively.
}
\examples{
\dontrun{
# Collect profile data.
target_file <- tempfile()
Rprof(target_file) # or start_profiler(target_file)
replicate(100, sessionInfo())
Rprof(NULL) # or stop_profiler()

# Convert to pprof format.
pprof_target_file <- tempfile()
profile_data <- profile::read_rprof(target_file)
profile::write_pprof(profile_data, pprof_target_file)

# Visualize profile data.
vis_pprof(pprof_target_file, port = 50001)

# Now, navigate to http://localhost:50001 in a web browser.

# Listen to remote connections.
vis_pprof(pprof_target_file, host = "0.0.0.0", port = 50002)

# Now, if the computer running vis_pprof() is available over your
# network, you can connect a different computer to the local
# pprof web server.
# E.g. http://domain-running-pprof.com:50002 or http:\\\\1.2.3.4:50002
# in a web browser.
}
}
