% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof.R
\name{find_pprof}
\alias{find_pprof}
\title{Find the pprof utility}
\usage{
find_pprof()
}
\description{
\code{pprof} is a tool for visualization and analysis of profiling data,
implemented in Go.
This function is a wrapper around \code{\link[proffer:find_pprof]{proffer::find_pprof()}}
for determining its location on your system.

This function will be called by \code{\link[=start_profiler]{start_profiler()}} to ensure
that profiling data can be processed by \code{\link[=stop_profiler]{stop_profiler()}}.
A functioning installation of \code{pprof} is a prerequisite to use
this package.
}
