% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler-package.R
\docType{package}
\name{jointprof-package}
\alias{jointprof}
\alias{jointprof-package}
\title{jointprof: Joint Profiling of Native C/C++ and R Code}
\description{
Simultaneously collects profiling data at both the
    R and C/C++ levels (using 'Rprof' and Google's 'libprofiler' from
    'gperftools') and collates the results to present a unified view of
    the collected data.  Supports writing these combined profiles to R's
    usual 'Rprof.out' format (for interoperability with existing tools) or
    interacting with the external 'pprof' program for its rich
    visualization tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-prof/jointprof}
  \item \url{http://r-prof.github.io/jointprof}
  \item Report bugs at \url{https://github.com/r-prof/jointprof/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item Google Inc. (Code adapted from gperftools under its BSD 3-clause license.) [copyright holder]
}

}
