% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{start_profiler}
\alias{start_profiler}
\alias{stop_profiler}
\title{Starts and stops profiling}
\usage{
start_profiler(path = "Rprof.out", ..., numfiles = 100L, bufsize = 10000L)

stop_profiler()
}
\arguments{
\item{path}{Path to the output file.}

\item{...}{Ignored, for extensibility.}

\item{numfiles, bufsize}{Passed on to \code{Rprof()} call.}
}
\value{
\code{stop_profiler()} returns the profiling data like it would have
been read by \code{\link[profile:read_rprof]{profile::read_rprof()}}.
}
\description{
\code{start_profiler()} initiates profiling. It is a replacement for \code{\link[=Rprof]{Rprof()}}
that will include native stack traces where available. Internally, \code{Rprof()}
is used to capture R call stacks.

\code{stop_profiler()} terminates profiling. The results are written to the
\code{Rprof()}-compatible file given specified by the \code{path} argument.
}
\details{
Profiling requires the \code{pprof} tool, which can be made available by
installing the \pkg{pprof} package (recommended) or by running
\code{go get github.com/google/pprof} and adding \code{${GOPATH}/bin} to the \code{PATH}.

Set the \code{keep.source} and \code{keep.source.pkgs} options to \code{TRUE} (via
\code{\link[=option]{option()}}) before installing packages from source or running code to obtain
accurate locations in your stack traces. It is a good idea to set these
options in your \link{.Rprofile} file.

Use the \pkg{profile} package to read the data or save the output in \code{pprof}
format for further processing.
}
\examples{
\dontrun{
start_profiler("Rprof.out")
# code to be profiled
stop_profiler()

profile::read_rprof("Rprof.out")
}
}
